Add: 등록한 음악정보 수정 기능

+++ client/src/Components/views/StreamsPage/StreamsPage.js
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
 import { getMusics, deleteMusic } from '../../../_actions/music_action';
 
 import classNames from 'classnames/bind';
@@ -15,8 +14,9 @@ function PlaylistsPage() {
     const user = useSelector(state => state.user.userData);
 
     const [Musics, setMusics] = useState([]);
-    const [Delete, setDelete] = useState(false);
-    const [Edit, setEdit] = useState(false);
+    const [MusicIndex, setMusicIndex] = useState(-1);
+    const [IsEdit, setIsEdit] = useState(false);
+    const [IsChange, setIsChange] = useState(false);
 
     useEffect(() => {
         dispatch(getMusics()).then(res => {
@@ -26,38 +26,41 @@ function PlaylistsPage() {
                 alert('Failed to get Musics');
             }
         });
-    }, [Delete]);
+        return setIsChange(false);
+    }, [IsChange]);
 
     const deleteCard = e => {
         const musicindex = e.target.getAttribute('musicindex');
 
-        let body = {
-            writer: user._id,
-            _id: Musics[musicindex]._id,
-        };
+        if (window.confirm('Want to delete this music?')) {
+            let body = {
+                writer: user._id,
+                _id: Musics[musicindex]._id,
+            };
 
-        dispatch(deleteMusic(body)).then(res => {
-            if (res.payload.success) {
-                // redux store의 state도 같이 변경시키자
-                setDelete(true);
-                // setMusics(Musics.splice(musicindex, 1));
-                // window.location.replace('/streams');
-            } else {
-                alert('Failed to delete music!');
-            }
-        });
+            dispatch(deleteMusic(body)).then(res => {
+                if (res.payload.success) {
+                    // redux store의 state도 같이 변경시키자
+                    setIsChange(true);
+                } else {
+                    alert('Failed to delete music!');
+                }
+            });
+        }
     };
 
-    const onClickEdit = () => {
-        setEdit(true);
+    const onClickEdit = e => {
+        const musicindex = e.target.getAttribute('musicindex');
+        setMusicIndex(musicindex);
+        setIsEdit(true);
     };
 
     const musicCards = Musics.map((music, index) => {
         return (
-            <Link to="/streams" className={cx('music')} key={index}>
+            <div className={cx('music')} key={index}>
                 <img
                     className={cx('image')}
-                    src="https://image.shutterstock.com/image-illustration/music-themed-thumbnail-260nw-1125825584.jpg"
+                    src="https://www.doc.gold.ac.uk/creativeprojects/wp-content/uploads/2016/10/unnamed.jpg"
                     alt="music"
                 />
                 <div className={cx('info')}>
@@ -68,7 +71,7 @@ function PlaylistsPage() {
                         <a className={cx('btn')} href={music.url} target="_blank" rel="noreferrer">
                             Link
                         </a>
-                        <button className={cx('btn')} onClick={onClickEdit}>
+                        <button className={cx('btn')} musicindex={index} onClick={onClickEdit}>
                             Edit
                         </button>
                         <button className={cx('btn')} musicindex={index} onClick={deleteCard}>
@@ -76,17 +79,17 @@ function PlaylistsPage() {
                         </button>
                     </div>
                 </div>
-            </Link>
+            </div>
         );
     });
 
     return (
-        <section className={Edit ? cx('wrapper', 'edit') : cx('wrapper')}>
+        <section className={cx('wrapper')}>
             <div className={cx('container')}>
                 <h1 className={cx('title')}>Musics</h1>
                 <div className={cx('musics')}>{musicCards}</div>
             </div>
-            {Edit && <EditMusicPage setEdit={setEdit} />}
+            {IsEdit && <EditMusicPage music={Musics[MusicIndex]} setIsEdit={setIsEdit} setIsChange={setIsChange} />}
         </section>
     );
 }
